requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
google-api-core==1.14.0   # via google-cloud-core
google-auth==1.6.3        # via google-api-core, google-cloud-storage
google-cloud-core==1.0.2  # via google-cloud-storage
google-cloud-storage==1.16.1
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
idna==2.8                 # via requests
protobuf==3.9.0           # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pytz==2019.1              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-auth, google-resumable-media, protobuf
urllib3==1.25.3           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via google-api-core, protobuf
